alternative linked list constructor
1. Clever implementation of public => boolean add(E element) to avoid duplicate code. Since add(int index, E element) is more generic, it can replace the simple add function.
2. I had to remove your throw new IndexOutOfBoundsException() from the helper function getNthNode(int index) so function add(int index, E element) works. The exception is now on each function. Some of them need: index >= size and other (add): index > size to work.
3. An alternative way to do the list constructor with fewer parameters. It also handles all the pointers for previous and next nodes.
4. A simple test at the end of class MyLinkedList<E> to see if constructor works.

I know this is just another way of doing stuff that yours was doing as fine. I used your code as inspiration but did a bit different for my assignment. Mostly aesthetic differences. The node constructor is now more dense and difficult to grasp, but it makes easier to understand/write the functions in the linked list class.
Feel free to accept or scrap these modifications. I just needed to know if I can read and understand other people code from something I did and modify them and check if they still worked.